+++ clients/pay/src/main/java/com/binance/connector/client/pay/rest/api/PayApi.java
@@ -42,7 +42,7 @@ public class PayApi {
 
     private static final String USER_AGENT =
             String.format(
-                    "binance-pay/1.1.0 (Java/%s; %s; %s)",
+                    "binance-pay/2.0.0 (Java/%s; %s; %s)",
                     SystemUtil.getJavaVersion(), SystemUtil.getOs(), SystemUtil.getArch());
     private static final boolean HAS_TIME_UNIT = false;
 
@@ -153,7 +153,11 @@ public class PayApi {
             localVarHeaderParams.put("Content-Type", localVarContentType);
         }
         List<String> localVarAuthNames = new ArrayList<>();
-        localVarAuthNames.addAll(Arrays.asList(new String[] {}));
+        localVarAuthNames.addAll(
+                Arrays.asList(
+                        new String[] {
+                            "binanceSignature",
+                        }));
         if (HAS_TIME_UNIT) {
             localVarAuthNames.add("timeUnit");
         }