+++ clients/derivatives-trading-portfolio-margin-pro/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 2.0.0 - 2025-05-26
+
+### Changed (1)
+
+- `queryPortfolioMarginProBankruptcyLoanRepayHistory()` (`GET /sapi/v1/portfolio/pmLoan-history` has been updated to `GET /sapi/v1/portfolio/pmloan-history`)
+- Marked as signed the following endpoints:
+    - `POST /sapi/v1/portfolio/repay`
+
 ## 1.1.0 - 2025-05-02
 
 - Update `binance/common` module to version `1.1.0`.