+++ clients/derivatives-trading-usds-futures/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## 2.0.0 - 2025-06-02
+### Removed (1)
+
+#### REST API
+
+- `historicalBlvtNavKlineCandlestick()` (`GET /fapi/v1/lvtKlines`)
+
+### Changed (1)
+
+#### REST API
+
+- Add `price` and `weight` to `constituents` object.
+
 ## 1.2.0 - 2025-05-13
 
 - Add proxy authentication for websocket