+++ clients/pay/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 2.0.0 - 2025-05-28
+
+### Changed (1)
+
+- Marked as signed the following endpoints:
+  - `GET /sapi/v1/pay/transactions`
+
 ## 1.1.0 - 2025-05-02
 
 - Update `binance/common` module to version `1.1.0`.